run custom model actions in one test to mitigate resouce usage
