+++ plugin/src/test/java/org/opensearch/ml/rest/RestMLCustomModelActionIT.java
@@ -8,22 +8,18 @@ package org.opensearch.ml.rest;
 import static org.opensearch.ml.common.MLTask.FUNCTION_NAME_FIELD;
 import static org.opensearch.ml.common.MLTask.MODEL_ID_FIELD;
 import static org.opensearch.ml.common.MLTask.STATE_FIELD;
-import static org.opensearch.ml.common.MLTask.TASK_ID_FIELD;
 
 import java.io.IOException;
 import java.util.Map;
 
-import org.apache.http.HttpEntity;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.rules.ExpectedException;
-import org.opensearch.client.Response;
-import org.opensearch.client.ResponseException;
 import org.opensearch.ml.common.MLTaskState;
 import org.opensearch.ml.common.transport.upload.MLUploadInput;
 import org.opensearch.ml.utils.TestHelper;
 
-public class RestMLLoadModelActionIT extends MLCommonsRestTestCase {
+public class RestMLCustomModelActionIT extends MLCommonsRestTestCase {
     @Rule
     public ExpectedException exceptionRule = ExpectedException.none();
 
@@ -34,23 +30,18 @@ public class RestMLLoadModelActionIT extends MLCommonsRestTestCase {
         uploadInput = createUploadModelInput();
     }
 
-    public void testLoadModelAPI_NoIndex() throws IOException {
-        exceptionRule.expect(ResponseException.class);
-        exceptionRule.expectMessage("no such index [.plugins-ml-model]");
-        loadModel("wrong_model_id");
-    }
-
-    public void testLoadModelAPI_Success() throws IOException, InterruptedException {
-        String uploadTaskId = uploadModel();
-        waitForTask(uploadTaskId, MLTaskState.COMPLETED);
-
-        getTask(client(), uploadTaskId, response -> {
+    public void testUnloadModelAPI_Success() throws IOException, InterruptedException {
+        // upload model
+        String taskId = uploadModel(TestHelper.toJsonString(uploadInput));
+        waitForTask(taskId, MLTaskState.COMPLETED);
+        getTask(client(), taskId, response -> {
             String algorithm = (String) response.get(FUNCTION_NAME_FIELD);
             assertEquals(uploadInput.getFunctionName().name(), algorithm);
             assertNotNull(response.get(MODEL_ID_FIELD));
             assertEquals(MLTaskState.COMPLETED.name(), response.get(STATE_FIELD));
             String modelId = (String) response.get(MODEL_ID_FIELD);
             try {
+                // load model
                 String loadTaskId = loadModel(modelId);
                 waitForTask(loadTaskId, MLTaskState.COMPLETED);
                 getTask(client(), loadTaskId, loadTaskResponse -> {
@@ -58,20 +49,14 @@ public class RestMLLoadModelActionIT extends MLCommonsRestTestCase {
                     assertEquals(MLTaskState.COMPLETED.name(), response.get(STATE_FIELD));
                 });
                 getModelProfile(modelId, verifyTextEmbeddingModelLoaded());
+                Map<String, Object> result = unloadModel(modelId);
+                for (Map.Entry<String, Object> entry : result.entrySet()) {
+                    Map stats = (Map) ((Map) entry.getValue()).get("stats");
+                    assertEquals("unloaded", stats.get(modelId));
+                }
             } catch (IOException | InterruptedException e) {
                 throw new RuntimeException(e);
             }
         });
-    }
-
-    private String uploadModel() throws IOException {
-        String input = TestHelper.toJsonString(uploadInput);
-        Response uploadResponse = TestHelper.makeRequest(client(), "POST", "/_plugins/_ml/models/_upload", null, input, null);
-        HttpEntity entity = uploadResponse.getEntity();
-        assertNotNull(uploadResponse);
-        String entityString = TestHelper.httpEntityToString(entity);
-        Map map = gson.fromJson(entityString, Map.class);
-        String taskId = (String) map.get(TASK_ID_FIELD);
-        return taskId;
     }
 }
